Feature/add token jwt

+++ doclyn-api/src/main/java/com/example/doclyn_api/services/FuncionarioService.java
@@ -28,6 +28,10 @@ public class FuncionarioService {
     }
 
     public void deleteById(Long id) {
+        if (!funcionarioRepository.existsById(id)) {
+            throw new ResourceNotFound("Nenhum funcionário com esse id foi encontrado!");
+        }
+
         funcionarioRepository.deleteById(id);
     }
